requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
attrs==25.1.0
backoff==2.2.1
boto3==1.37.4
botocore==1.37.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
filelock==3.17.0
filetype==1.2.0
fire==0.7.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
huggingface-hub==0.24.7
idna==3.7
inference-sdk==0.40.0
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.8
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.54
s3transfer==0.11.3
sahi==0.11.20
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
shapely==2.0.7
six==1.17.0
smmap==5.0.2
supervision==0.25.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.81
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
yolov5==7.0.14
